renamed and moved some functions and directories.
