+++ dashboard/Instructions.md
@@ -1,0 +1,3 @@
+from ./dashboard run: `bokeh serve --show app`
+
+Run `pre_processing_data`manually after any GETTSIM parameter is changed.
