+++ dashboard/app/plots/deductions.py
@@ -4,7 +4,7 @@ from bokeh.models import Div
 from bokeh.models import Panel
 from bokeh.palettes import Category20
 from bokeh.plotting import figure
-from Scripts.plotstyle import plotstyle
+from plots.plotstyle import plotstyle
 
 
 def deductions(plot_dict, data):
