+++ dashboard/app/plots/plotstyle.py
