+++ dashboard/app/plots/social_assistance.py
@@ -4,17 +4,17 @@ from bokeh.models import Div
 from bokeh.models import Panel
 from bokeh.palettes import Category20
 from bokeh.plotting import figure
-from Scripts.plotstyle import plotstyle
+from plots.plotstyle import plotstyle
 
 
-def social_security(plot_dict, data):
+def social_assistance(plot_dict, data):
     def setup_plot(src):
-
         p = figure(
-            plot_width=900,
+            plot_width=750,
             plot_height=400,
-            x_range=(1984, 2022 + 8),
-            tooltips="$name: @$name{0.00%}",
+            y_range=(0, 500),
+            x_range=(2005, 2022),
+            tooltips="$name: @$name €",
         )
 
         labels = list(src.data.keys())[1:]
@@ -48,12 +48,23 @@ def social_security(plot_dict, data):
                 line_color=colors[i],
                 name=labels[i],
             )
-        p.xaxis.bounds = (1984, max(src.data["index"]) + 1)
+        p.xaxis.bounds = (2004, max(src.data["index"]) + 1)
+        p.xaxis.ticker.desired_num_ticks = 4
 
         plot = plotstyle(p, plot_dict)
 
         return plot
 
+    data = data.rename(
+        columns={
+            "ein_erwachsener": "Single Adult",
+            "zwei_erwachsene": "Adults in Couple",
+            "weitere_erwachsene": "Adults not in Couple",
+            "kinder_14_24": "Child 14 to 17 years",
+            "kinder_7_13": "Child 6 to 13 years",
+            "kinder_0_6": "Child < 6 years",
+        }
+    )
     src = ColumnDataSource(data)
 
     p = setup_plot(src)
@@ -61,6 +72,6 @@ def social_security(plot_dict, data):
 
     layout = column(description, p)
 
-    tab = Panel(child=layout, title="Social security contributions")
+    tab = Panel(child=layout, title="Social Assistance Rate")
 
     return tab
